Reader: removed traces of old-post
