+++ client/reader/stream/index.jsx
@@ -9,9 +9,7 @@ import { defer, flatMap, lastIndexOf, noop, times, clamp } from 'lodash';
 /**
  * Internal dependencies
  */
-import config from 'config';
 import ReaderMain from 'components/reader-main';
-import Main from 'components/main';
 import DISPLAY_TYPES from 'lib/feed-post-store/display-types';
 import EmptyContent from './empty';
 import * as FeedStreamStoreActions from 'lib/feed-stream-store/actions';
@@ -21,9 +19,8 @@ import LikeStoreActions from 'lib/like-store/actions';
 import LikeHelper from 'reader/like-helper';
 import InfiniteList from 'components/infinite-list';
 import MobileBackToSidebar from 'components/mobile-back-to-sidebar';
-import Post from './post';
 import CrossPost from './x-post';
-import RefreshPost from './refresh-post';
+import Post from './post';
 import page from 'page';
 import PostPlaceholder from './post-placeholder';
 import PostStore from 'lib/feed-post-store';
@@ -39,7 +36,7 @@ import FeedSubscriptionStore from 'lib/reader-feed-subscriptions';
 const GUESSED_POST_HEIGHT = 600;
 const HEADER_OFFSET_TOP = 46;
 
-function oldCardFactory( post ) {
+function cardFactory( post ) {
 	if ( post.display_type & DISPLAY_TYPES.X_POST ) {
 		return CrossPost;
 	}
@@ -50,16 +47,6 @@ function oldCardFactory( post ) {
 
 	return Post;
 }
-
-function refreshCardFactory( post ) {
-	let postClass = oldCardFactory( post );
-	if ( postClass === Post ) {
-		postClass = RefreshPost;
-	}
-	return postClass;
-}
-
-const defaultCardFactory = config.isEnabled( 'reader/refresh/stream' ) ? refreshCardFactory : oldCardFactory;
 
 const MIN_DISTANCE_BETWEEN_RECS = 4; // page size is 7, so one in the middle of every page and one on page boundries, sometimes
 const MAX_DISTANCE_BETWEEN_RECS = 30;
@@ -200,7 +187,7 @@ export default class ReaderStream extends React.Component {
 				return externalPostClass;
 			}
 		}
-		return defaultCardFactory( post );
+		return cardFactory( post );
 	}
 
 	scrollToSelectedPost( animate ) {
@@ -541,10 +528,8 @@ export default class ReaderStream extends React.Component {
 			showingStream = true;
 		}
 
-		const CurrentMain = config.isEnabled( 'reader/refresh/stream' ) ? ReaderMain : Main;
-
 		return (
-			<CurrentMain className={ classnames( 'following', this.props.className ) }>
+			<ReaderMain className={ classnames( 'following', this.props.className ) }>
 				{ this.props.showMobileBackToSidebar && <MobileBackToSidebar>
 					<h1>{ this.props.listName }</h1>
 				</MobileBackToSidebar> }
@@ -556,7 +541,7 @@ export default class ReaderStream extends React.Component {
 					? <div className="infinite-scroll-end" />
 					: null
 				}
-			</CurrentMain>
+			</ReaderMain>
 		);
 	}
 }
