+++ client/reader/stream/post.jsx
@@ -1,411 +1,137 @@
 /**
- * External Dependencies
+ * External dependencies
  */
-const ReactDom = require( 'react-dom' ),
-	React = require( 'react' ),
-	PureRenderMixin = require( 'react-pure-render/mixin' ),
-	assign = require( 'lodash/assign' ),
-	classnames = require( 'classnames' ),
-	closest = require( 'component-closest' ),
-	//debug = require( 'debug' )( 'calypso:reader:following:post' ),
-	head = require( 'lodash/head' ),
-	filter = require( 'lodash/filter' ),
-	forOwn = require( 'lodash/forOwn' ),
-	startsWith = require( 'lodash/startsWith' ),
-	twemoji = require( 'twemoji' ),
-	get = require( 'lodash/get' ),
-	page = require( 'page' ),
-	pick = require( 'lodash/pick' );
+import React from 'react';
+import { connect } from 'react-redux';
+import { get } from 'lodash';
 
 /**
- * Internal Dependencies
+ * Internal dependencies
  */
-const
-	Card = require( 'components/card' ),
-	CommentButton = require( 'blocks/comment-button' ),
-	DISPLAY_TYPES = require( 'lib/feed-post-store/display-types' ),
-	EmbedContainer = require( 'components/embed-container' ),
-	LikeButton = require( 'reader/like-button' ),
-	ObserveWindowSizeMixin = require( 'lib/mixins/observe-window-resize' ),
-	PostHeader = require( 'reader/post-header' ),
-	PostByline = require( 'reader/post-byline' ),
-	PostImages = require( 'reader/post-images' ),
-	PostOptions = require( 'reader/post-options' ),
-	PostErrors = require( 'reader/post-errors' ),
-	PostExcerpt = require( 'components/post-excerpt' ),
-	SiteStore = require( 'lib/reader-site-store' ),
-	SiteStoreActions = require( 'lib/reader-site-store/actions' ),
-	Share = require( 'reader/share' ),
-	ShareHelper = require( 'reader/share/helper' ),
-	utils = require( 'reader/utils' ),
-	PostCommentHelper = require( 'reader/comments/helper' ),
-	LikeHelper = require( 'reader/like-helper' ),
-	readerRoute = require( 'reader/route' ),
-	stats = require( 'reader/stats' ),
-	PostPermalink = require( 'reader/post-permalink' ),
-	DailyPostHelper = require( 'blocks/daily-post-button/helper' ),
-	DailyPostButton = require( 'blocks/daily-post-button' ),
-	DiscoverPostAttribution = require( 'reader/discover/post-attribution' ),
-	DiscoverSiteAttribution = require( 'reader/discover/site-attribution' ),
-	DiscoverHelper = require( 'reader/discover/helper' ),
-	FeedPostStore = require( 'lib/feed-post-store' ),
-	FeedPostStoreActions = require( 'lib/feed-post-store/actions' ),
-	Gridicon = require( 'components/gridicon' ),
-	smartSetState = require( 'lib/react-smart-set-state' ),
-	FeaturedAsset = require( 'reader/stream/featured-asset' );
+import ReaderPostCard from 'blocks/reader-post-card';
+import { getSite } from 'state/reader/sites/selectors';
+import { getFeed } from 'state/reader/feeds/selectors';
+import QueryReaderSite from 'components/data/query-reader-site';
+import QueryReaderFeed from 'components/data/query-reader-feed';
+import FeedPostStore from 'lib/feed-post-store';
+import smartSetState from 'lib/react-smart-set-state';
+import { recordAction, recordGaEvent, recordTrackForPost } from 'reader/stats';
+import {
+	isDiscoverPost,
+	isDiscoverSitePick,
+	getSourceData as getDiscoverSourceData,
+	discoverBlogId
+ } from 'reader/discover/helper';
 
-const Post = React.createClass( {
+class ReaderPostCardAdapter extends React.Component {
 
-	mixins: [ PureRenderMixin, ObserveWindowSizeMixin ],
-
-	propTypes: {
-		post: React.PropTypes.object.isRequired,
-		isSelected: React.PropTypes.bool.isRequired,
-		xPostedTo: React.PropTypes.array,
-		suppressSiteNameLink: React.PropTypes.bool,
-		showPostHeader: React.PropTypes.bool,
-		showFollowInHeader: React.PropTypes.bool,
-		additionalClasses: React.PropTypes.object,
-		handleClick: React.PropTypes.func.isRequired,
-	},
-
-	smartSetState: smartSetState,
-
-	shouldApplyIsLong: function() {
-		var node = ReactDom.findDOMNode( this.refs.siteName );
-		// give the clientWidth a 2 pixel buffer. IE is often off by at least one.
-		return !! ( node && node.scrollWidth > ( node.offsetWidth + 2 ) );
-	},
-
-	onWindowResize: function() {
-		this.checkSiteNameForOverflow();
-	},
-
-	getInitialState: function() {
-		return this.getStateFromStores();
-	},
-
-	getDefaultProps: function() {
-		return {
-			showFollowInHeader: false
-		};
-	},
-
-	getStateFromStores: function( props ) {
-		var post, siteId, site, originalPost, isDiscoverPost, isDiscoverSitePick;
-		props = props || this.props;
-		post = props.post;
-		siteId = ! post.is_external && post.site_ID;
-		isDiscoverPost = post && DiscoverHelper.isDiscoverPost( post );
-		if ( isDiscoverPost ) {
-			isDiscoverSitePick = post && DiscoverHelper.isDiscoverSitePick( post );
-		}
-
-		if ( siteId ) {
-			site = SiteStore.get( siteId );
-			if ( ! site ) {
-				SiteStoreActions.fetch( siteId );
-			}
-		}
-
-		// If it's a discover post (but not a site pick), we want the original post too
-		if ( isDiscoverPost && ! isDiscoverSitePick && get( post, 'discover_metadata.featured_post_wpcom_data.blog_id' ) ) {
-			originalPost = FeedPostStore.get( {
-				blogId: post.discover_metadata.featured_post_wpcom_data.blog_id,
-				postId: post.discover_metadata.featured_post_wpcom_data.post_id
-			} );
-		}
-
-		return {
-			site: site,
-			siteish: utils.siteishFromSiteAndPost( site, post ),
-			originalPost: originalPost,
-			isDiscoverPost: isDiscoverPost,
-			isDiscoverSitePick: isDiscoverSitePick
-		};
-	},
-
-	componentDidMount: function() {
-		this.checkSiteNameForOverflow();
-		this._parseEmoji();
-		SiteStore.on( 'change', this.updateState );
-	},
-
-	componentWillUnmount: function() {
-		SiteStore.off( 'change', this.updateState );
-	},
-
-	componentWillReceiveProps: function( nextProps ) {
-		if ( nextProps.post !== this.props.post ) {
-			this.updateState( nextProps );
-		}
-	},
-
-	updateState: function( props ) {
-		var newState = this.getStateFromStores( props );
-		if ( newState.site !== this.state.site ) {
-			this.smartSetState( newState );
-		}
-	},
-
-	componentDidUpdate: function() {
-		this.checkSiteNameForOverflow();
-		this._parseEmoji();
-	},
-
-	featuredImageComponent: function( post ) {
-		var featuredImage = ( post.canonical_image && pick( post.canonical_image, [ 'uri', 'width', 'height' ] ) ),
-			featuredEmbed = head( filter( post.content_embeds, ( embed ) => {
-				return ! startsWith( embed.type, 'special-' );
-			} ) ),
-			useFeaturedEmbed;
-
-		if ( post.use_excerpt ) {
-			// don't feature embeds for excerpts
-			featuredEmbed = null;
-		}
-		if ( ! ( featuredImage || featuredEmbed ) ) {
-			return null;
-		}
-
-		// we only show a featured embed when all of these are true
-		//   - there is no featured image on the post that's big enough to pass as the canonical image
-		//   - there is an available embed
-		//
-		useFeaturedEmbed = featuredEmbed &&
-			( ! featuredImage || ( featuredImage !== post.featured_image && featuredImage !== get( post, 'post_thumbnail.URL' ) ) );
-
-		const featuredAssetProps = {
-			featuredImage,
-			featuredEmbed,
-			useFeaturedEmbed
-		};
-		return <FeaturedAsset { ...featuredAssetProps } />;
-	},
-
-	checkSiteNameForOverflow: function() {
-		var headerNode = ReactDom.findDOMNode( this.refs.siteName );
-		if ( ! headerNode ) {
-			return;
-		}
-		headerNode.classList.toggle( 'is-long', this.shouldApplyIsLong() );
-	},
-
-	propagateCardClick: function( options = {} ) {
-		let postToOpen = this.props.post;
-		// For Discover posts (but not site picks), open the original post in full post view
-		if ( this.state.originalPost ) {
-			postToOpen = {
-				...this.state.originalPost,
+	onClick = ( postToOpen ) => {
+		let referredPost;
+		if ( this.props.originalPost && isDiscoverPost( this.props.post ) && ! isDiscoverSitePick( this.props.post ) ) {
+			referredPost = { ...postToOpen,
 				referral: {
-					blogId: DiscoverHelper.discoverBlogId,
-					postId: this.props.ID
+					blogId: discoverBlogId,
+					postId: this.props.post.ID
 				}
 			};
 		}
-
-		this.props.handleClick( postToOpen, options );
-	},
-
-	handleCardClick: function( event ) {
-		var rootNode = ReactDom.findDOMNode( this ),
-			selection = window.getSelection && window.getSelection();
-
-		// if the click has modifier or was not primary, ignore it
-		if ( event.button > 0 || event.metaKey || event.controlKey || event.shiftKey || event.altKey ) {
-			if ( closest( event.target, '.reader__post-title-link', true, rootNode ) ) {
-				stats.recordPermalinkClick( 'card_title_with_modifier', this.props.post );
-			}
-			return;
-		}
-
-		if ( closest( event.target, '.should-scroll', true, rootNode ) ) {
-			setTimeout( function() {
-				window.scrollTo( 0, 0 );
-			}, 100 );
-		}
-
-		// declarative ignore
-		if ( closest( event.target, '.ignore-click, [rel~=external]', true, rootNode ) ) {
-			return;
-		}
-
-		// ignore clicks on anchors inside inline content
-		if ( closest( event.target, 'a', true, rootNode ) && closest( event.target, '.reader-full-post__story-content', true, rootNode ) ) {
-			return;
-		}
-
-		// ignore clicks when highlighting text
-		if ( selection && selection.toString() ) {
-			return;
-		}
-
-		// programattic ignore
-		if ( ! event.defaultPrevented ) { // some child handled it
-			event.preventDefault();
-			this.propagateCardClick();
-		}
-	},
-
-	handleCommentButtonClick: function() {
-		stats.recordAction( 'click_comments' );
-		stats.recordGaEvent( 'Clicked Post Comment Button' );
-		stats.recordTrackForPost( 'calypso_reader_post_comments_button_clicked', this.props.post );
-		this.propagateCardClick( { comments: true } );
-	},
-
-	_parseEmoji: function() {
-		twemoji.parse( ReactDom.findDOMNode( this ) );
-	},
-
-	pickSite: function( event ) {
-		// ugh, double negative. If we should let the site click go, bail.
-		if ( utils.isSpecialClick( event ) ) {
-			return;
-		}
-
-		const url = readerRoute.getStreamUrlFromPost( this.props.post );
-		page.show( url );
-	},
-
-	handleSiteClick: function( event ) {
-		if ( ! utils.isSpecialClick( event ) ) {
-			event.preventDefault();
-		}
-	},
-
-	maybeMarkPostSeen: function( liked ) {
-		if ( liked ) {
-			FeedPostStoreActions.markSeen( this.props.post );
-		}
-	},
-
-	render: function() {
-		var post = this.props.post,
-			site = this.state.siteish,
-			featuredImage = this.featuredImageComponent( post ),
-			shouldShowComments = PostCommentHelper.shouldShowComments( post ),
-			shouldShowLikes = LikeHelper.shouldShowLikes( post ),
-			shouldShowShare = ShareHelper.shouldShowShare( post ),
-			hasFeaturedImage = featuredImage !== null,
-			articleClasses = assign( {
-				reader__card: true,
-				'has-featured-image': hasFeaturedImage,
-				'hide-xpost': this.props.xPostedTo
-			}, this.additionalClasses ),
-			shouldShowExcerptOnly = !! post.use_excerpt,
-			shouldUseFullExcerpt = ! shouldShowExcerptOnly && ( post.display_type & DISPLAY_TYPES.ONE_LINER ),
-			siteName = utils.siteNameFromSiteAndPost( this.state.site, post ),
-			isDiscoverPost = this.state.isDiscoverPost,
-			isDiscoverSitePick = this.state.isDiscoverSitePick,
-			discoverSiteUrl,
-			originalPost = this.state.originalPost,
-			likeSiteId = ( originalPost ? originalPost.site_ID : post.site_ID ),
-			likePostId = ( originalPost ? originalPost.ID : post.ID ),
-			commentCount = ( originalPost ? originalPost.discussion.comment_count : post.discussion.comment_count ),
-			xPostedToContent = null;
-
-		forOwn( DISPLAY_TYPES, function( value, key ) {
-			if ( post.display_type && ( post.display_type & value ) ) { // bitwise intentional
-				articleClasses[ 'is-' + key.toLowerCase().replace( '_', '-' ) ] = true;
-			}
-		} );
-
-		if ( shouldShowExcerptOnly ) {
-			articleClasses[ 'is-excerpt-only' ] = true;
-		}
-
-		if ( this.props.isSelected ) {
-			articleClasses[ 'is-selected' ] = true;
-		}
-
-		if ( post.site_ID ) {
-			articleClasses[ 'blog-' + post.site_ID ] = true;
-		}
-
-		if ( post.feed_ID ) {
-			articleClasses[ 'feed-' + post.feed_ID ] = true;
-		}
-
-		if ( ! this.props.showPostHeader ) {
-			articleClasses[ 'is-headerless' ] = true;
-		}
-
-		forOwn( post.tags, ( { slug } ) => {
-			articleClasses[ 'tag-' + slug ] = true;
-		} );
-
-		articleClasses = classnames( articleClasses );
-
-		if ( isDiscoverPost ) {
-			discoverSiteUrl = DiscoverHelper.getSiteUrl( post );
-		}
-
-		if ( this.props.xPostedTo ) {
-			xPostedToContent = this.props.xPostedTo.map( ( xPostedTo, index, array ) => {
-				return (
-					<span className="reader__x-post-site" key={ xPostedTo.siteURL + '-' + index } >
-						<a href={ xPostedTo.siteURL }>{ xPostedTo.siteName }</a>
-						{ index + 2 < array.length ? <span>, </span> : null }
-						{ index + 2 === array.length ? <span> { this.translate( 'and', { comment: 'last conjuction in a list of blognames: (blog1, blog2,) blog3 _and_ blog4' } ) } </span> : null }
-					</span>
-				);
-			} );
-		}
-
-		return (
-			<Card tagName="article" onClick={ this.handleCardClick } className={ articleClasses }>
-
-				<PostErrors post={ post } />
-
-				{ this.props.showPostHeader ? <PostHeader site={ site } siteUrl={ post.site_URL } showFollow={ this.props.showFollowInHeader } onSiteSelect={ this.pickSite } onSiteClick={ this.handleSiteClick } /> : null }
-
-				{ featuredImage }
-
-				{ post.title ? <h1 className="reader__post-title"><a className="reader__post-title-link" href={ post.URL } target="_blank" rel="noopener noreferrer">{ post.title }</a></h1> : null }
-
-				<PostByline post={ post } site={ this.props.site } isDiscoverPost={ isDiscoverPost } />
-
-				{ shouldUseFullExcerpt
-					? <EmbedContainer>
-							<div key="full-post-inline"
-								className="reader-full-post__story-content"
-								dangerouslySetInnerHTML={ // eslint-disable-line react/no-danger
-									{ __html: post.content }
-								} />
-						</EmbedContainer>
-					: <PostExcerpt content={ post.better_excerpt ? post.better_excerpt : post.excerpt } />
-				}
-
-				{ ! shouldShowExcerptOnly
-					? <PostImages postImages={ post.content_images || [] } />
-					: null }
-
-				{ ( isDiscoverPost && post.discover_metadata && ! isDiscoverSitePick ) ? <DiscoverPostAttribution attribution={ post.discover_metadata.attribution } siteUrl={ discoverSiteUrl } followUrl={ DiscoverHelper.getSourceFollowUrl( post ) } /> : null }
-				{ ( isDiscoverSitePick ) ? <DiscoverSiteAttribution attribution={ post.discover_metadata.attribution } siteUrl={ discoverSiteUrl } followUrl={ DiscoverHelper.getSourceFollowUrl( post ) } /> : null }
-
-				{ this.props.xPostedTo
-					? <div className="reader__x-post-to">
-							<Gridicon icon="arrow-right" size={ 24 } />
-							<span className="reader__x-post-label">{ this.translate( 'cross-posted to' ) } </span>
-							{ xPostedToContent }
-						</div>
-					: null
-				}
-
-				{ DailyPostHelper.isDailyPostChallengeOrPrompt( post ) ? <DailyPostButton post={ post } tagName="span" /> : null }
-
-				<ul className="reader__post-footer">
-					<PostPermalink siteName={ siteName } postUrl={ post.URL } />
-					{ ( shouldShowShare ) ? <Share post={ post } /> : null }
-					{ ( shouldShowComments ) ? <CommentButton onClick={ this.handleCommentButtonClick } commentCount={ commentCount } /> : null }
-					{ ( shouldShowLikes ) ? <LikeButton siteId={ likeSiteId } postId={ likePostId } onLikeToggle={ this.maybeMarkPostSeen } /> : null }
-					<li className="reader__post-options"><PostOptions post={ post } site={ this.state.site } /></li>
-				</ul>
-			</Card>
-		);
+		this.props.handleClick && this.props.handleClick( referredPost || postToOpen );
 	}
 
-} );
+	onCommentClick = () => {
+		recordAction( 'click_comments' );
+		recordGaEvent( 'Clicked Post Comment Button' );
+		recordTrackForPost( 'calypso_reader_post_comments_button_clicked', this.props.post );
 
-module.exports = Post;
+		this.props.handleClick && this.props.handleClick( this.props.post, { comments: true } );
+	}
+
+	// take what the stream hands to a card and adapt it
+	// for use by a ReaderPostCard
+	render() {
+		const {
+			feed_ID: feedId,
+			site_ID: siteId,
+			is_external: isExternal
+		} = this.props.post;
+		const _isDiscoverPost = isDiscoverPost( this.props.post );
+
+		// only query the site if the feed id is missing. feed queries end up fetching site info
+		// via a meta query, so we don't need both.
+		return (
+			<ReaderPostCard
+				post={ this.props.post }
+				originalPost={ this.props.originalPost }
+				site={ this.props.site }
+				feed={ this.props.feed }
+				onClick={ this.onClick }
+				onCommentClick={ this.onCommentClick }
+				isSelected={ this.props.isSelected }
+				showPrimaryFollowButton={ this.props.showPrimaryFollowButtonOnCards }
+				showEntireExcerpt={ _isDiscoverPost }
+				useBetterExcerpt={ ! _isDiscoverPost }
+				showSiteName={ this.props.showSiteName }>
+				{ feedId && <QueryReaderFeed feedId={ feedId } includeMeta={ false } /> }
+				{ ! isExternal && siteId && <QueryReaderSite siteId={ +siteId } includeMeta={ false } /> }
+			</ReaderPostCard>
+		);
+	}
+}
+
+const ConnectedReaderPostCardAdapter = connect(
+	( state, ownProps ) => {
+		const siteId = get( ownProps, 'post.site_ID' );
+		const isExternal = get( ownProps, 'post.is_external' );
+		const feedId = get( ownProps, 'post.feed_ID' );
+		return {
+			site: isExternal ? null : getSite( state, siteId ),
+			feed: getFeed( state, feedId ),
+			originalPost: ownProps.originalPost
+		};
+	}
+)( ReaderPostCardAdapter );
+
+/**
+ * A container for the ReaderPostCardAdapter responsible for binding to Flux stores
+ */
+export default class ReaderPostCardAdapterFluxContainer extends React.Component {
+	constructor( props ) {
+		super( props );
+		this.state = this.getStateFromStores( props );
+		this.smartSetState = smartSetState;
+	}
+
+	getStateFromStores( props = this.props ) {
+		const post = props.post;
+		const nonSiteDiscoverPick = isDiscoverPost( post ) && ! isDiscoverSitePick( post );
+
+		// If it's a discover post (but not a site pick), we want the original post too
+		const originalPost = nonSiteDiscoverPick ? FeedPostStore.get( getDiscoverSourceData( post ) ) : null;
+
+		return {
+			originalPost
+		};
+	}
+
+	updateState = ( newState = this.getStateFromStores() ) => {
+		this.smartSetState( newState );
+	}
+
+	componentWillMount() {
+		FeedPostStore.on( 'change', this.updateState );
+	}
+
+	componentWillReceiveProps( nextProps ) {
+		this.updateState( this.getStateFromStores( nextProps ) );
+	}
+
+	componentWillUnmount() {
+		FeedPostStore.off( 'change', this.updateState );
+	}
+
+	render() {
+		return ( <ConnectedReaderPostCardAdapter
+					{ ...this.props }
+					originalPost={ this.state.originalPost } /> );
+	}
+}
